Ajna methods interfaces clean up
