+++ packages/dma-library/src/protocols/ajna/index.ts
@@ -1,33 +1,20 @@
 import poolAbi from '@abis/external/protocols/ajna/ajnaPoolERC20.json'
-import { Address } from '@deploy-configurations/types/address'
 import { ZERO } from '@dma-common/constants'
 import { negativeToZero } from '@dma-common/utils/common'
 import { getAjnaEarnValidations } from '@dma-library/strategies/ajna/earn/validations'
 import { getPoolLiquidity } from '@dma-library/strategies/ajna/validation/notEnoughLiquidity'
-import { AjnaEarnPosition } from '@dma-library/types/ajna'
-import { AjnaPool } from '@dma-library/types/ajna/ajna-pool'
 import {
-  AjnaDependencies,
+  AjnaCommonDependencies,
   AjnaEarnActions,
+  AjnaEarnPosition,
   AjnaError,
   AjnaWarning,
   Strategy,
-} from '@dma-library/types/common'
+} from '@dma-library/types/ajna'
+import { AjnaEarnPayload } from '@dma-library/types/ajna/ajna-dependencies'
+import { AjnaPool } from '@dma-library/types/ajna/ajna-pool'
 import BigNumber from 'bignumber.js'
 import { ethers } from 'ethers'
-
-export interface AjnaEarnArgs {
-  poolAddress: Address
-  dpmProxyAddress: Address
-  collateralAmount: BigNumber
-  quoteAmount: BigNumber
-  quoteTokenPrecision: number
-  price: BigNumber
-  position: AjnaEarnPosition
-  collateralPrice: BigNumber
-  quotePrice: BigNumber
-  isStakingNft?: boolean
-}
 
 export const prepareAjnaPayload = <T extends { pool: AjnaPool }>({
   dependencies,
@@ -37,7 +24,7 @@ export const prepareAjnaPayload = <T extends { pool: AjnaPool }>({
   data,
   txValue,
 }: {
-  dependencies: AjnaDependencies
+  dependencies: AjnaCommonDependencies
   targetPosition: T
   errors: AjnaError[]
   warnings: AjnaWarning[]
@@ -70,8 +57,8 @@ export const getAjnaEarnActionOutput = async ({
 }: {
   targetPosition: AjnaEarnPosition
   data: string
-  dependencies: AjnaDependencies
-  args: AjnaEarnArgs
+  dependencies: AjnaCommonDependencies
+  args: AjnaEarnPayload
   action: AjnaEarnActions
   txValue: string
 }) => {
