+++ packages/dma-library/src/strategies/ajna/earn/claim-collateral.ts
@@ -1,16 +1,16 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 import ajnaProxyActionsAbi from '@abis/external/protocols/ajna/ajnaProxyActions.json'
-import { AjnaEarnArgs, getAjnaEarnActionOutput } from '@dma-library/protocols/ajna'
-import { AjnaEarnPosition } from '@dma-library/types/ajna'
-import { AjnaDependencies, Strategy } from '@dma-library/types/common'
+import { getAjnaEarnActionOutput } from '@dma-library/protocols/ajna'
+import { AjnaCommonDependencies, AjnaEarnPosition, Strategy } from '@dma-library/types/ajna'
+import { AjnaEarnPayload } from '@dma-library/types/ajna/ajna-dependencies'
 import BigNumber from 'bignumber.js'
 import { ethers } from 'ethers'
 
 import bucketPrices from './buckets.json'
 
 export type AjnaClaimCollateralStrategy = (
-  args: AjnaEarnArgs,
-  dependencies: AjnaDependencies,
+  args: AjnaEarnPayload,
+  dependencies: AjnaCommonDependencies,
 ) => Promise<Strategy<AjnaEarnPosition>>
 
 export const claimCollateral: AjnaClaimCollateralStrategy = async (args, dependencies) => {
