+++ packages/dma-library/src/strategies/ajna/earn/deposit-adjust.ts
@@ -2,21 +2,17 @@
 import ajnaProxyActionsAbi from '@abis/external/protocols/ajna/ajnaProxyActions.json'
 import poolInfoAbi from '@abis/external/protocols/ajna/poolInfoUtils.json'
 import { ZERO } from '@dma-common/constants'
-import {
-  AjnaEarnArgs,
-  getAjnaEarnActionOutput,
-  resolveAjnaEthAction,
-} from '@dma-library/protocols/ajna'
-import { AjnaEarnPosition } from '@dma-library/types/ajna'
-import { AjnaDependencies, Strategy } from '@dma-library/types/common'
+import { getAjnaEarnActionOutput, resolveAjnaEthAction } from '@dma-library/protocols/ajna'
+import { AjnaCommonDependencies, AjnaEarnPosition, Strategy } from '@dma-library/types/ajna'
+import { AjnaEarnPayload } from '@dma-library/types/ajna/ajna-dependencies'
 import BigNumber from 'bignumber.js'
 import { ethers } from 'ethers'
 
 import bucketPrices from './buckets.json'
 
 export type AjnaDepositAndAdjustStrategy = (
-  args: AjnaEarnArgs,
-  dependencies: AjnaDependencies,
+  args: AjnaEarnPayload,
+  dependencies: AjnaCommonDependencies,
 ) => Promise<Strategy<AjnaEarnPosition>>
 
 export const depositAndAdjust: AjnaDepositAndAdjustStrategy = async (args, dependencies) => {
