+++ packages/dma-library/src/strategies/ajna/earn/withdraw-adjust.ts
@@ -2,17 +2,17 @@
 import ajnaProxyActionsAbi from '@abis/external/protocols/ajna/ajnaProxyActions.json'
 import poolInfoAbi from '@abis/external/protocols/ajna/poolInfoUtils.json'
 import { ZERO } from '@dma-common/constants'
-import { AjnaEarnArgs, getAjnaEarnActionOutput } from '@dma-library/protocols/ajna'
-import { AjnaEarnPosition } from '@dma-library/types/ajna'
-import { AjnaDependencies, Strategy } from '@dma-library/types/common'
+import { getAjnaEarnActionOutput } from '@dma-library/protocols/ajna'
+import { AjnaCommonDependencies, AjnaEarnPosition, Strategy } from '@dma-library/types/ajna'
+import { AjnaEarnPayload } from '@dma-library/types/ajna/ajna-dependencies'
 import BigNumber from 'bignumber.js'
 import { ethers } from 'ethers'
 
 import bucketPrices from './buckets.json'
 
 export type AjnaWithdrawAndAdjustStrategy = (
-  args: AjnaEarnArgs,
-  dependencies: AjnaDependencies,
+  args: AjnaEarnPayload,
+  dependencies: AjnaCommonDependencies,
 ) => Promise<Strategy<AjnaEarnPosition>>
 
 export const withdrawAndAdjust: AjnaWithdrawAndAdjustStrategy = async (args, dependencies) => {
