+++ packages/dma-library/src/strategies/ajna/index.ts
@@ -1,4 +1,6 @@
-import { AjnaDepositBorrowStrategy, depositBorrow } from './deposit-borrow'
+import { AjnaDepositBorrowStrategy, depositBorrow } from './borrow/deposit-borrow'
+import { AjnaOpenBorrowStrategy, open as openBorrow } from './borrow/open'
+import { AjnaPaybackWithdrawStrategy, paybackWithdraw } from './borrow/payback-withdraw'
 import { AjnaClaimCollateralStrategy, claimCollateral } from './earn/claim-collateral'
 import { AjnaDepositAndAdjustStrategy, depositAndAdjust } from './earn/deposit-adjust'
 import { AjnaOpenEarnStrategy, open as openEarn } from './earn/open'
@@ -6,8 +8,6 @@ import { AjnaWithdrawAndAdjustStrategy, withdrawAndAdjust } from './earn/withdra
 import { adjustMultiply, AjnaAdjustRiskStrategy } from './multiply/adjust'
 import { AjnaCloseStrategy, closeMultiply } from './multiply/close'
 import { AjnaOpenMultiplyStrategy, openMultiply } from './multiply/open'
-import { AjnaOpenBorrowStrategy, open as openBorrow } from './open'
-import { AjnaPaybackWithdrawStrategy, paybackWithdraw } from './payback-withdraw'
 
 export const ajna: {
   borrow: {
