+++ packages/dma-library/src/strategies/ajna/multiply/open.ts
@@ -1,40 +1,20 @@
-import { Address } from '@deploy-configurations/types/address'
 import { prepareAjnaPayload, resolveAjnaEthAction } from '@dma-library/protocols/ajna'
-import { AjnaPosition } from '@dma-library/types/ajna'
-import { Strategy } from '@dma-library/types/common'
+import {
+  AjnaCommonDependencies,
+  AjnaOpenMultiplyPayload,
+  AjnaPosition,
+  Strategy,
+} from '@dma-library/types/ajna'
 import { views } from '@dma-library/views'
-import { GetPoolData } from '@dma-library/views/ajna'
-import { IRiskRatio } from '@domain/risk-ratio'
 import BigNumber from 'bignumber.js'
-import { ethers } from 'ethers'
-
-export interface OpenMultiplyArgs {
-  poolAddress: Address
-  dpmProxyAddress: Address
-  collateralPrice: BigNumber
-  quotePrice: BigNumber
-  quoteTokenPrecision: number
-  collateralAmount: BigNumber
-  collateralTokenPrecision: number
-  riskRatio: IRiskRatio
-}
-
-export interface Dependencies {
-  poolInfoAddress: Address
-  ajnaProxyActions: Address
-  provider: ethers.providers.Provider
-  WETH: Address
-  getPoolData: GetPoolData
-  getPosition?: typeof views.ajna.getPosition
-}
 
 export type AjnaOpenMultiplyStrategy = (
-  args: OpenMultiplyArgs,
-  dependencies: Dependencies,
+  args: AjnaOpenMultiplyPayload,
+  dependencies: AjnaCommonDependencies,
 ) => Promise<Strategy<AjnaPosition>>
 
 export const openMultiply: AjnaOpenMultiplyStrategy = async (args, dependencies) => {
-  const getPosition = dependencies.getPosition ? dependencies.getPosition : views.ajna.getPosition
+  const getPosition = views.ajna.getPosition
   const position = await getPosition(
     {
       collateralPrice: args.collateralPrice,
