+++ packages/dma-library/src/types/ajna/ajna-dependencies.ts
@@ -1,0 +1,65 @@
+import { Address } from '@deploy-configurations/types/address'
+import { AjnaEarnPosition, AjnaPosition } from '@dma-library/types'
+import { GetEarnData } from '@dma-library/views'
+import { GetPoolData } from '@dma-library/views/ajna'
+import { IRiskRatio } from '@domain'
+import BigNumber from 'bignumber.js'
+import { ethers } from 'ethers'
+
+export interface AjnaCommonDependencies {
+  ajnaProxyActions: Address
+  poolInfoAddress: Address
+  provider: ethers.providers.Provider
+  WETH: Address
+  getPoolData: GetPoolData
+}
+
+export interface AjnaOpenEarnDependencies extends AjnaCommonDependencies {
+  getEarnData: GetEarnData
+  rewardsManagerAddress: Address
+}
+
+export interface AjnaCommonPayload {
+  dpmProxyAddress: string
+  poolAddress: string
+  collateralPrice: BigNumber
+  collateralTokenPrecision: number
+  quotePrice: BigNumber
+  quoteTokenPrecision: number
+}
+
+export interface AjnaOpenBorrowPayload extends AjnaCommonPayload {
+  quoteAmount: BigNumber
+  collateralAmount: BigNumber
+}
+
+export interface AjnaBorrowPayload extends AjnaCommonPayload {
+  position: AjnaPosition
+  quoteAmount: BigNumber
+  collateralAmount: BigNumber
+}
+
+export interface AjnaOpenEarnPayload extends AjnaCommonPayload {
+  isStakingNft: boolean
+  price: BigNumber
+  quoteAmount: BigNumber
+}
+
+export interface AjnaOpenMultiplyPayload extends AjnaCommonPayload {
+  collateralAmount: BigNumber
+  riskRatio: IRiskRatio
+}
+
+export interface AjnaMultiplyPayload extends AjnaCommonPayload {
+  position: AjnaPosition
+  collateralAmount: BigNumber
+  riskRatio: IRiskRatio
+}
+
+export interface AjnaEarnPayload extends AjnaCommonPayload {
+  isStakingNft: boolean
+  price: BigNumber
+  quoteAmount: BigNumber
+  collateralAmount: BigNumber
+  position: AjnaEarnPosition
+}
