+++ packages/dma-library/src/types/ajna/ajna-strategy.ts
@@ -1,0 +1,14 @@
+import { Tx } from '@dma-common/types'
+import { AjnaError, AjnaWarning } from '@dma-library/types/ajna'
+
+export type Strategy<Position> = {
+  simulation: {
+    swaps: []
+    // @deprecated - use position
+    targetPosition: Position
+    position: Position
+    errors: AjnaError[]
+    warnings: AjnaWarning[]
+  }
+  tx: Tx
+}
