+++ packages/dma-library/src/types/ajna/index.ts
@@ -1,2 +1,27 @@
+export type {
+  AjnaBorrowPayload,
+  AjnaCommonDependencies,
+  AjnaCommonPayload,
+  AjnaMultiplyPayload,
+  AjnaOpenEarnDependencies,
+  AjnaOpenEarnPayload,
+  AjnaOpenMultiplyPayload,
+} from './ajna-dependencies'
+export type { AjnaEarnActions } from './ajna-earn-position'
 export { AjnaEarnPosition } from './ajna-earn-position'
 export { AjnaPosition } from './ajna-position'
+export type { Strategy } from './ajna-strategy'
+export type {
+  AjnaError,
+  AjnaErrorAfterLupIndexBiggerThanHtpIndex,
+  AjnaErrorBorrowUndercollateralized,
+  AjnaErrorDustLimit,
+  AjnaErrorNotEnoughLiquidity,
+  AjnaErrorOverRepay,
+  AjnaErrorOverWithdraw,
+  AjnaErrorWithdrawMoreThanAvailable,
+  AjnaErrorWithdrawUndercollateralized,
+  AjnaWarning,
+  AjnaWarningPriceAboveMomp,
+  AjnaWarningPriceBelowHtp,
+} from './ajna-validations'
